Webtoon: Use repository patern
